gcp_postgres_sandbox/utilities/organized_structure/requirements.txt
# Core dependencies for CryptoPrism Database Utilities
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
sqlalchemy>=2.0.32
psycopg2-binary>=2.9.0
requests>=2.32.3
typing-extensions>=4.12.2
python-dateutil>=2.8.0

# Optional visualization dependencies (install with: pip install -r requirements-viz.txt)
# matplotlib>=3.7.0
# seaborn>=0.12.0
# sqlalchemy-schemadisplay>=1.3
# graphviz>=0.20.0

# Optional AI features (install with: pip install -r requirements-ai.txt)  
# google-generativeai>=0.5.0

# Optional notification features (install with: pip install -r requirements-notifications.txt)
# python-telegram-bot>=13.15

# Optional MySQL support (install with: pip install -r requirements-mysql.txt)
# mysql-connector-python>=9.0.0